core/json: Rename json_operations.rs to ops.rs
